Added events handling and it works as I expect it to work. It alerts you when you press BACK or you don't have internet. Changed the thing that happens when you get hit, now if the game state is GAME_OVER or the hitNumber is zero it doesn't do anything. I might wanna change it so that it checks if the hitNumber has already been sent, but for now this works. If the received playerList is zero then it keeps the old one. This is due to a bug I was having where it'd fuck up and it would send an empty list when you sent hits after the game was over. Gonna start adding map stuff now.
Added map script to index.html file. Organized buttons to show up in the top bar or as elements of the lists. Now it looks cleaner and nicer.

Updated the target from 14 to 15, because it was screwing up. And BluetoothPlugin.java says we require API 15, so it must be so!